refactor: 현장톡을 livetalk으로 통일화
