+++ android/app/src/main/java/com/yagubogu/presentation/livetalk/LiveTalkFragment.kt
@@ -1,15 +1,15 @@
-package com.yagubogu.presentation.talk
+package com.yagubogu.presentation.livetalk
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.yagubogu.databinding.FragmentTalkBinding
+import com.yagubogu.databinding.FragmentLiveTalkBinding
 
 @Suppress("ktlint:standard:backing-property-naming")
-class TalkFragment : Fragment() {
-    private var _binding: FragmentTalkBinding? = null
+class LiveTalkFragment : Fragment() {
+    private var _binding: FragmentLiveTalkBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -17,7 +17,7 @@ class TalkFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        _binding = FragmentTalkBinding.inflate(inflater, container, false)
+        _binding = FragmentLiveTalkBinding.inflate(inflater, container, false)
         return binding.root
     }
