+++ android/app/src/main/res/drawable/ic_livetalk.xml
