+++ android/app/src/main/res/layout/fragment_live_talk.xml
@@ -5,7 +5,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.talk.TalkFragment">
+        tools:context=".presentation.livetalk.LiveTalkFragment">
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
